collapse struct context into coexpression block
The context struct has been expanded with many fields and is not maintained
in the cstate anymore. Most fields belong in either the coexpression block
or the thread state. All fields are migrated to the the coexpression block
for now. Some of them will be moved to the thread state later. This simplifies
maintaining the state of a coexpression/thread.